// pronti-frontend-main/front/src/app/modulos/sri-gastos/form/form-sri-gastos.enum.ts
export enum FormSriGastosEnum {
  // nombre = 'nombre',
  anio = 'anio',
  gastoVivienda = 'gastoVivienda',
  gastoEducacion = 'gastoEducacion',
  gastoSalud = 'gastoSalud',
  gastoVestido = 'gastoVestido',
  gastoAlimento = 'gastoAlimento',
  gastoTurismo = 'gastoTurismo',
  discapacidad = 'discapacidad',
  terceraEdad = 'terceraEdad',
  totalGastos = 'totalGastos',
  idTrabajador = 'idTrabajador'
}